# backend/docs/swagger.yaml
basePath: /
definitions:
...
  do.BindRequest:
    properties:
      channel_id:
        type: string
      code:
        type: string
    type: object
  do.BindingDTO:
    properties:
      channel_id:
        example: "0"
        type: string
      channel_name:
        type: string
      channel_type:
        type: string
      created_at:
        type: string
      id:
        example: "0"
        type: string
      platform_user_id:
        type: string
      user_id:
        example: "0"
        type: string
    type: object
...
  do.CreatePushEventRequest:
    properties:
      channels:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      event_key:
        type: string
      targets:
        items:
          type: string
        type: array
      task_type:
        type: string
      template:
        type: string
    type: object
...
  do.PublicChannelDTO:
    properties:
      id:
        example: "0"
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  do.TestPushRequest:
    properties:
      config:
        $ref: '#/definitions/push.Config'
      target:
        type: string
    required:
    - config
    type: object
...
  do.UpdatePushEventRequest:
    properties:
      channels:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      targets:
        items:
          type: string
        type: array
      template:
        type: string
    required:
    - template
    type: object
...
  entity.PushEvent:
    properties:
      channels:
        items:
          type: string
        type: array
      created_at:
        type: string
      enabled:
        type: boolean
      event_key:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      targets:
        items:
          type: string
        type: array
      task_type:
        type: string
      template:
        type: string
      updated_at:
        type: string
    type: object
...
  push.Config:
    properties:
      channel:
        description: 渠道名称，例如 "lark", "custom", "email" 等，唯一标识
        type: string
      ext:
        additionalProperties: {}
        description: 预留拓展 JSON 配置
        type: object
      key:
        description: AppID 或 SMTP 用户名
        type: string
      other:
        description: 附加配置 (如 ChatID / UserKey / 扩展 JSON)
        type: string
      secret:
        description: 签名密钥或 SMTP 密码/Token
        type: string
      url:
        description: Webhook 地址或 SMTP 地址
        type: string
    type: object
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
info:
  contact:
    name: Wavelet
    url: https://github.com/Rain-kl/Wavelet
  description: Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Wavelet API
  version: 1.0.0
paths:
...
  /api/v1/admin/push/events:
    get:
      description: 返回系统配置的通知事件列表，包括预置和自定义事件，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 通知事件列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.PushEvent'
                  type: array
              type: object
      security:
      - SessionCookie: []
      summary: 获取所有通知事件
      tags:
      - admin-push
    post:
      consumes:
      - application/json
      description: 绑定系统内置事件或异步任务、推送渠道、接收目标并创建通知事件配置，需要管理员权限
      parameters:
      - description: 创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.CreatePushEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/entity.PushEvent'
              type: object
      security:
      - SessionCookie: []
      summary: 创建通知事件
      tags:
      - admin-push
  /api/v1/admin/push/events/{id}:
    delete:
      description: 删除数据库中的特定通知事件配置，需要管理员权限
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  type: string
              type: object
      security:
      - SessionCookie: []
      summary: 删除通知事件配置
      tags:
      - admin-push
    put:
      consumes:
      - application/json
      description: 更新已有通知事件的推送渠道、接收目标和内容模板，需要管理员权限
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.UpdatePushEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  type: string
              type: object
      security:
      - SessionCookie: []
      summary: 更新通知事件
      tags:
      - admin-push
  /api/v1/admin/push/events/{id}/toggle:
    post:
      description: 启用或禁用指定的通知事件
      parameters:
      - description: 事件 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 切换成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  type: string
              type: object
      security:
      - SessionCookie: []
      summary: 快捷切换通知事件启用状态
      tags:
      - admin-push
  /api/v1/admin/push/events/builtin:
    get:
      description: 返回系统定义的所有内置通知事件元数据，供前端下拉框选择，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 内置通知事件列表
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取所有内置通知事件
      tags:
      - admin-push
  /api/v1/admin/push/histories:
    get:
      description: 返回分页的通知历史日志数据，需要管理员权限
      parameters:
      - description: 当前页码
        in: query
        name: page
        type: integer
      - description: 分页大小
        in: query
        name: page_size
        type: integer
      - description: 过滤事件名称
        in: query
        name: event_key
        type: string
      - description: 过滤发送状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 推送历史列表
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 分页获取通知推送历史
      tags:
      - admin-push
  /api/v1/admin/push/test:
    post:
      consumes:
      - application/json
      description: 接收临时通知渠道配置并在本地同步调用 Pusher.Send 发送测试消息
      parameters:
      - description: 测试请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.TestPushRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 测试成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  type: string
              type: object
      security:
      - SessionCookie: []
      summary: 测试推送通道发送
      tags:
      - admin-push
...
  /api/v1/message-gateway/bindings:
    get:
      description: Returns the current user's bound messaging channels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/do.BindingDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: List message gateway bindings
      tags:
      - message-gateway
    post:
      consumes:
      - application/json
      description: Binds the current user to a platform identity using a one-time
        pairing code
      parameters:
      - description: bind body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.BindRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/do.BindingDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: Bind a messaging channel
      tags:
      - message-gateway
  /api/v1/message-gateway/bindings/{id}:
    delete:
      description: Removes a binding owned by the current user
      parameters:
      - description: binding id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: Unbind a messaging channel
      tags:
      - message-gateway
  /api/v1/message-gateway/channels:
    get:
      description: Returns enabled system bots the current user can pair with
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/do.PublicChannelDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: List enabled messaging channels
      tags:
      - message-gateway
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"